Don't load application code before application is initialized
