+++ sentry-rails/spec/dummy/test_rails_app/app.rb
@@ -23,18 +23,22 @@ v6_1 = Gem::Version.new("6.1")
 v7_0 = Gem::Version.new("7.0")
 v7_1 = Gem::Version.new("7.1")
 
-case Gem::Version.new(Rails.version)
-when -> (v) { v < v5_2 }
-  require "dummy/test_rails_app/apps/5-0"
-when -> (v) { v.between?(v5_2, v6_0) }
-  require "dummy/test_rails_app/apps/5-2"
-when -> (v) { v.between?(v6_0, v6_1) }
-  require "dummy/test_rails_app/apps/6-0"
-when -> (v) { v.between?(v6_1, v7_0) }
-  require "dummy/test_rails_app/apps/6-1"
-when -> (v) { v.between?(v7_0, v7_1) }
-  require "dummy/test_rails_app/apps/7-0"
-end
+FILE_NAME =
+  case Gem::Version.new(Rails.version)
+  when -> (v) { v < v5_2 }
+    "5-0"
+  when -> (v) { v.between?(v5_2, v6_0) }
+    "5-2"
+  when -> (v) { v.between?(v6_0, v6_1) }
+    "6-0"
+  when -> (v) { v.between?(v6_1, v7_0) }
+    "6-1"
+  when -> (v) { v.between?(v7_0, v7_1) }
+    "7-0"
+  end
+
+# require files and defined relevant setup methods for the Rails version
+require "dummy/test_rails_app/configs/#{FILE_NAME}"
 
 def make_basic_app(&block)
   run_pre_initialize_cleanup
@@ -85,6 +89,9 @@ def make_basic_app(&block)
   app.initialize!
 
   Rails.application = app
+
+  # load application code for the Rails version
+  require "dummy/test_rails_app/apps/#{FILE_NAME}"
 
   Post.all.to_a # to run the sqlte version query first
