+++ sentry-rails/spec/dummy/test_rails_app/apps/6-1.rb
@@ -1,6 +1,3 @@
-require "active_storage/engine"
-require "action_cable/engine"
-
 ActiveRecord::Schema.define do
   create_table "active_storage_attachments", force: :cascade do |t|
     t.string "name", null: false
@@ -119,13 +116,3 @@ class HelloController < ActionController::Base
     raise ActionController::BadRequest
   end
 end
-
-def run_pre_initialize_cleanup
-  ActionCable::Channel::Base.reset_callbacks(:subscribe)
-  ActionCable::Channel::Base.reset_callbacks(:unsubscribe)
-end
-
-def configure_app(app)
-  app.config.active_storage.service = :test
-end
-
