+++ sentry-rails/spec/dummy/test_rails_app/configs/5-0.rb
@@ -1,0 +1,3 @@
+def run_pre_initialize_cleanup; end
+
+def configure_app(app); end
