+++ sentry-rails/spec/dummy/test_rails_app/configs/5-2.rb
@@ -1,0 +1,8 @@
+require "active_storage/engine"
+
+def run_pre_initialize_cleanup; end
+
+def configure_app(app)
+  app.config.active_storage.service = :test
+end
+
